E-com website clone
This project is a fully responsive e-commerce website built using HTML, CSS, and JavaScript. It showcases a modern design and provides essential features for an online store, including product listings, a shopping cart, and a checkout process. The website is designed to offer a seamless shopping experience on both desktop and mobile devices.